+++ domain/src/main/java/ru/dpopkov/knownics/domain/BaseEntity.java
@@ -1,7 +1,14 @@
 package ru.dpopkov.knownics.domain;
 
-public abstract class BaseEntity {
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.MappedSuperclass;
 
+@MappedSuperclass
+public abstract class BaseEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     public Long getId() {