+++ repository-jpa/src/main/java/ru/dpopkov/knownics/repository/jpa/Profiles.java
@@ -1,0 +1,5 @@
+package ru.dpopkov.knownics.repository.jpa;
+
+public class Profiles {
+    public static final String REPOSITORY_JPA = "repository-jpa";
+}