+++ repository-map/src/main/java/ru/dpopkov/knownics/repository/map/DeckMapRepository.java
@@ -1,10 +1,12 @@
 package ru.dpopkov.knownics.repository.map;
 
+import org.springframework.context.annotation.Profile;
 import ru.dpopkov.knownics.domain.deck.Deck;
 import ru.dpopkov.knownics.domain.deck.DeckRepository;
 import ru.dpopkov.knownics.domain.deck.FlashCard;
 import ru.dpopkov.knownics.domain.deck.FlashCardRepository;
 
+@Profile(Profiles.REPOSITORY_MAP)
 public class DeckMapRepository extends AbstractMapRepository<Deck> implements DeckRepository {
 
     private final FlashCardRepository flashCardRepository;