+++ repository-map/src/main/java/ru/dpopkov/knownics/repository/map/KeyTermMapRepository.java
@@ -1,6 +1,8 @@
 package ru.dpopkov.knownics.repository.map;
 
+import org.springframework.context.annotation.Profile;
 import ru.dpopkov.knownics.domain.KeyTerm;
 
+@Profile(Profiles.REPOSITORY_MAP)
 public class KeyTermMapRepository extends AbstractMapRepository<KeyTerm> {
 }