+++ repository-map/src/main/java/ru/dpopkov/knownics/repository/map/Profiles.java
@@ -1,0 +1,5 @@
+package ru.dpopkov.knownics.repository.map;
+
+public class Profiles {
+    public static final String REPOSITORY_MAP = "repository-map";
+}