+++ repository-map/src/main/java/ru/dpopkov/knownics/repository/map/SourceDetailsMapRepository.java
@@ -1,10 +1,12 @@
 package ru.dpopkov.knownics.repository.map;
 
+import org.springframework.context.annotation.Profile;
 import ru.dpopkov.knownics.domain.answer.Source;
 import ru.dpopkov.knownics.domain.answer.SourceDetails;
 import ru.dpopkov.knownics.domain.answer.SourceDetailsRepository;
 import ru.dpopkov.knownics.domain.answer.SourceRepository;
 
+@Profile(Profiles.REPOSITORY_MAP)
 public class SourceDetailsMapRepository extends AbstractMapRepository<SourceDetails> implements SourceDetailsRepository {
 
     private final SourceRepository sourceRepository;