+++ web/src/main/java/ru/dpopkov/knownics/Profiles.java
@@ -1,0 +1,6 @@
+package ru.dpopkov.knownics;
+
+public class Profiles {
+    public static final String REPOSITORY_MAP = ru.dpopkov.knownics.repository.map.Profiles.REPOSITORY_MAP;
+    public static final String REPOSITORY_JPA = ru.dpopkov.knownics.repository.jpa.Profiles.REPOSITORY_JPA;
+}